arch: migrate redteam runtimes to surface shims and align surface policy roots
