+++ client/runtime/config/root_surface_contract.json
@@ -47,6 +47,7 @@
     "packages",
     "planes",
     "proofs",
+    "surface",
     "tests",
     "tools"
   ],
